requirements.txt
streamlit>=1.32.0
supabase>=2.3.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
nltk>=3.8.0
numpy>=1.26.0
plotly>=5.0.0